requirements.txt
discord.py==2.5.2
python-dotenv==1.1.0
yt-dlp==2025.3.31
PyNaCl==1.5.0
Flask==3.1.0